add documentation and update README

+++ install.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 
+function echoerr () {
+    echo -e $* 1>&2
+}
+
 echo -e "...Affect Installation...\n"
 
 if ! command -v opam >/dev/null ; then
-  echo "You are missing opam, the OCaml package manager."
-  echo "Visit the following link for help installing it:"
-  echo "  https://opam.ocaml.org/doc/Install.html"
+  echoerr "You are missing opam, the OCaml package manager."
+  echoerr "Visit the following link for help installing it:"
+  echoerr "  https://opam.ocaml.org/doc/Install.html"
   exit 1
 fi
 
@@ -14,10 +18,10 @@ echo -e "> This can take a few minutes..."
 opam update
 
 if [ -d _opam ] ; then
-  echo "There is already a local opam switch in the current directory."
-  echo "If it is OK to remove it, please type:"
-  echo "  opam switch remove ."
-  echo "then run ./install.sh again."
+  echoerr "There is already a local opam switch in the current directory."
+  echoerr "If it is OK to remove it, please type:"
+  echoerr "  opam switch remove ."
+  echoerr "then run ./install.sh again."
   exit 1
 fi
 
@@ -25,6 +29,12 @@ export OPAMYES=true
 
 echo -e "\n> Creating a new local opam switch."
 echo -e "> This will take a while (perhaps over 10 minutes)..."
+
+if [[ `opam switch show 2> /dev/null` == 'affect' ]]; then
+    echoerr -e "\n'affect' switch already installed. Please remove switch (opam switch remove affect) to reinstall"
+    exit 1
+fi
+
 opam switch create affect \
   --no-install \
   --repositories=default,coq-released=https://coq.inria.fr/opam/released,iris-dev=git+https://gitlab.mpi-sws.org/iris/opam.git \
@@ -38,10 +48,15 @@ opam update
 opam pin add -n coq 8.18.0
 
 echo -e "\n> Fetching coq-hazel library\n"
+if [[ -d hazel ]]; then
+    echoerr "'hazel' directory already present. Please remove it before continuing"
+    exit 1
+fi
+
 git clone https://gitlab.inria.fr/cambium/hazel
 if (( $? != 0)); then
-    echo "Unable to fetch coq-hazel from https://gitlab.inria.fr/cambium/hazel"
-    echo "Using local copy instead."
+    echoerr "Unable to fetch coq-hazel from https://gitlab.inria.fr/cambium/hazel"
+    echoerr "Using local copy instead."
     [[ -d hazel ]] && rm -r hazel
     tar xzf hazel-local-copy.tar.gz
 fi
